+++ src/main/resources/application.properties
@@ -1,0 +1,16 @@
+# change port
+server.port = 7777
+# dbcp setting 
+spring.datasource.url=jdbc:oracle:thin:@127.0.0.1:1521:xe
+spring.datasource.username=scott
+spring.datasource.password=tiger
+# springmvc viewResolver setting 
+spring.mvc.view.prefix=/WEB-INF/views/
+spring.mvc.view.suffix=.jsp
+#devtools
+spring.devtools.livereload.enabled=true
+# mybatis
+mybatis.type-aliases-package=org.kosta.myproject.model.domain
+mybatis.configuration.map-underscore-to-camel-case=true
+# log level setting
+logging.level.root=ERROR
