+++ src/main/resources/org/kosta/myproject/model/mapper/MemberMapper.xml
@@ -15,6 +15,12 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 		select count(*) from g_member
 	</select>
 	
+	<update id="updateMember">
+		update g_member set name=#{name}, password=#{password}, address=#{address}, phonenum=#{phonenum}, 
+		birth=to_date(#{birth},'YYYY-MM-DD') where email=#{email}
+	</update>
+	
+	
 	<insert id="registerMember" parameterType="MemberDTO">
 		insert into 
 		g_member(member_id,email,name,nickname,address,phonenum,password,birth)
