+++ src/main/webapp/WEB-INF/views/member/update_result.jsp
@@ -4,9 +4,10 @@
  <%-- spring security custom tag를 사용하기 위한 선언 --%>
 <%@taglib prefix="sec"  uri="http://www.springframework.org/security/tags"%>
 <%-- javascript 한글처리가 안될 때는 변수에 담아서 el로 처리한다  --%>
-<sec:authentication property="principal.name" var="memberName"/>
+<sec:authentication property="principal" var="member"/>
  <script>
- alert('${memberName}님 회원정보수정되었습니다.');
+ alert(${member.name}+"님 회원정보수정되었습니다.");
+ location.href="/";
  </script>
  </body>
 </html>
